Support Immortalis-G925, Mali-G725, and Mali-G625
This change also updates to latest counter naming and equation derivations.

+++ backend/device/src/device/hwcnt/sampler/timestamp.hpp
@@ -55,7 +55,7 @@ class timestamp_iface {
 #endif
         constexpr uint64_t nsec_per_sec = 1000000000;
 
-        return now.tv_sec * nsec_per_sec + now.tv_nsec;
+        return static_cast<uint64_t>(now.tv_sec) * nsec_per_sec + static_cast<uint64_t>(now.tv_nsec);
     }
 };
